test: isolating test runs
- kicking off new process for each test file
- pulling out two problematic tests into separate files to run in isolation
