+++ package.json
@@ -20,7 +20,7 @@
     "node": ">=0.8"
   },
   "scripts": {
-    "test": "mocha --expose-gc --slow 2000 --timeout 600000"
+    "test": "set -e; for t in test/*.js; do mocha --expose-gc --slow 2000 --timeout 600000 $t; done"
   },
   "keywords": ["zeromq", "zmq", "0mq", "ømq", "libzmq", "native", "binding", "addon"],
   "license": "MIT",
