+++ test/socket.monitor.js
@@ -55,52 +55,4 @@ describe('socket.monitor', function() {
       });
     });
   });
-
-  it('should use default interval and numOfEvents', function(done) {
-    var req = zmq.socket('req');
-    req.setsockopt(zmq.ZMQ_RECONNECT_IVL, 5); // We want a quick connect retry from zmq
-
-    // We will try to connect to a non-existing server, zmq will issue events: "connect_retry", "close", "connect_retry"
-    // The connect_retry will be issued immediately after the close event, so we will measure the time between the close
-    // event and connect_retry event, those should >= 10 (this will tell us that we are reading 1 event at a time from
-    // the monitor socket).
-
-    var closeTime;
-    req.on('close', function() {
-	    closeTime = Date.now();
-    });
-
-    req.on('connect_retry', function() {
-      var diff = Date.now() - closeTime;
-      req.unmonitor();
-      req.close();
-      diff.should.be.within(10, 20);
-      done();
-    });
-
-    req.monitor();
-    req.connect('tcp://127.0.0.1:5423');
-  });
-
-  it('should read multiple events on monitor interval', function(done) {
-    var req = zmq.socket('req');
-    req.setsockopt(zmq.ZMQ_RECONNECT_IVL, 5);
-    var closeTime;
-    req.on('close', function() {
-      closeTime = Date.now();
-    });
-
-    req.on('connect_retry', function() {
-      var diff = Date.now() - closeTime;
-      req.unmonitor();
-      req.close();
-      diff.should.be.within(0, 5);
-      done();
-    });
-
-    // This should read all available messages from the queue, and we expect that "close" and "connect_retry" will be
-    // read on the same interval (for further details see the comment in the previous test)
-    req.monitor(10, 0);
-    req.connect('tcp://127.0.0.1:5423');
-  });
 });
